Moved the console-env tests into appropriate new testfile
